requirements.txt
aiofiles==24.1.0
alembic==1.15.1
black==25.1.0
faker==37.0.0
fastapi==0.115.11
flytekit==1.15.3
isort==6.0.1
mypy==1.15.0
numpy>=1.23.0
pandas>=1.5.0
psycopg2-binary>=2.9.0
pydantic==2.10.6
pydantic_settings==2.8.1
python-dotenv>=0.19.0
python-multipart==0.0.20
pytest==8.3.5
sqlalchemy>=1.4.0
strawberry-graphql==0.262.5
tqdm==4.67.1
uvicorn==0.34.0
